# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "BadgeProgress": {
        "properties": {
          "badge_id": {
            "format": "uuid",
            "title": "Badge Id",
            "type": "string"
          },
          "category": {
            "title": "Category",
            "type": "string"
          },
          "code": {
            "title": "Code",
            "type": "string"
          },
          "credit_value": {
            "title": "Credit Value",
            "type": "integer"
          },
          "current_value": {
            "title": "Current Value",
            "type": "integer"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "earned": {
            "title": "Earned",
            "type": "boolean"
          },
          "earned_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Earned At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "progress_percent": {
            "title": "Progress Percent",
            "type": "number"
          },
          "threshold": {
            "title": "Threshold",
            "type": "integer"
          }
        },
        "required": [
          "badge_id",
          "code",
          "name",
          "description",
          "category",
          "threshold",
          "current_value",
          "earned",
          "earned_at",
          "progress_percent",
          "credit_value"
        ],
        "title": "BadgeProgress",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RedemptionAction": {
        "properties": {
          "parent_notes": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Notes"
          },
          "status": {
            "enum": [
              "approved",
              "rejected",
              "fulfilled",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status"
        ],
        "title": "RedemptionAction",
        "type": "object"
      },
      "RedemptionCreate": {
        "properties": {
          "reader_id": {
            "format": "uuid",
            "title": "Reader Id",
            "type": "string"
          },
          "reward_item_id": {
            "format": "uuid",
            "title": "Reward Item Id",
            "type": "string"
          }
        },
        "required": [
          "reader_id",
          "reward_item_id"
        ],
        "title": "RedemptionCreate",
        "type": "object"
      },
      "RedemptionResponse": {
        "properties": {
          "approved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved At"
          },
          "cancelled_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cancelled At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "credit_cost": {
            "title": "Credit Cost",
            "type": "integer"
          },
          "fulfilled_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fulfilled At"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "parent_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Notes"
          },
          "reader_id": {
            "format": "uuid",
            "title": "Reader Id",
            "type": "string"
          },
          "requested_at": {
            "format": "date-time",
            "title": "Requested At",
            "type": "string"
          },
          "reward_item_id": {
            "format": "uuid",
            "title": "Reward Item Id",
            "type": "string"
          },
          "reward_name": {
            "title": "Reward Name",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "reader_id",
          "reward_item_id",
          "reward_name",
          "credit_cost",
          "status",
          "requested_at",
          "approved_at",
          "fulfilled_at",
          "cancelled_at",
          "parent_notes",
          "created_at",
          "updated_at"
        ],
        "title": "RedemptionResponse",
        "type": "object"
      },
...
      "RewardItemCreate": {
        "properties": {
          "active": {
            "default": true,
            "title": "Active",
            "type": "boolean"
          },
          "credit_cost": {
            "exclusiveMinimum": 0.0,
            "maximum": 1000.0,
            "title": "Credit Cost",
            "type": "integer"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "image_url": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "quantity": {
            "anyOf": [
              {
                "maximum": 10000.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          }
        },
        "required": [
          "name",
          "credit_cost"
        ],
        "title": "RewardItemCreate",
        "type": "object"
      },
      "RewardItemResponse": {
        "properties": {
          "active": {
            "title": "Active",
            "type": "boolean"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "credit_cost": {
            "title": "Credit Cost",
            "type": "integer"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "household_id": {
            "format": "uuid",
            "title": "Household Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "quantity": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "household_id",
          "name",
          "description",
          "credit_cost",
          "image_url",
          "quantity",
          "active",
          "created_at",
          "updated_at"
        ],
        "title": "RewardItemResponse",
        "type": "object"
      },
      "RewardItemUpdate": {
        "properties": {
          "active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active"
          },
          "credit_cost": {
            "anyOf": [
              {
                "exclusiveMinimum": 0.0,
                "maximum": 1000.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credit Cost"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 1000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "image_url": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "quantity": {
            "anyOf": [
              {
                "maximum": 10000.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quantity"
          }
        },
        "title": "RewardItemUpdate",
        "type": "object"
      },
      "RewardProgressResponse": {
        "properties": {
          "badges": {
            "items": {
              "$ref": "#/components/schemas/BadgeProgress"
            },
            "title": "Badges",
            "type": "array"
          },
          "credit_balance": {
            "title": "Credit Balance",
            "type": "integer"
          },
          "current_continuous_days": {
            "title": "Current Continuous Days",
            "type": "integer"
          },
          "current_week_reading_days": {
            "title": "Current Week Reading Days",
            "type": "integer"
          },
          "current_weekly_streak": {
            "title": "Current Weekly Streak",
            "type": "integer"
          },
          "finished_books": {
            "title": "Finished Books",
            "type": "integer"
          },
          "longest_continuous_days": {
            "title": "Longest Continuous Days",
            "type": "integer"
          },
          "longest_weekly_streak": {
            "title": "Longest Weekly Streak",
            "type": "integer"
          },
          "reader_id": {
            "format": "uuid",
            "title": "Reader Id",
            "type": "string"
          }
        },
        "required": [
          "reader_id",
          "credit_balance",
          "finished_books",
          "current_week_reading_days",
          "current_weekly_streak",
          "longest_weekly_streak",
          "current_continuous_days",
          "longest_continuous_days",
          "badges"
        ],
        "title": "RewardProgressResponse",
        "type": "object"
      },
      "RewardTransactionResponse": {
        "properties": {
          "amount": {
            "title": "Amount",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "reader_id": {
            "format": "uuid",
            "title": "Reader Id",
            "type": "string"
          },
          "source_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "transaction_type": {
            "title": "Transaction Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "reader_id",
          "amount",
          "transaction_type",
          "source_id",
          "description",
          "created_at"
        ],
        "title": "RewardTransactionResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "My Reading Tracker API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/reward-items": {
      "get": {
        "operationId": "list_reward_items_api_v1_reward_items_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_inactive",
            "required": false,
            "schema": {
              "default": true,
              "title": "Include Inactive",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RewardItemResponse"
                  },
                  "title": "Response List Reward Items Api V1 Reward Items Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Reward Items",
        "tags": [
          "rewards"
        ]
      },
      "post": {
        "operationId": "create_reward_item_api_v1_reward_items_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RewardItemCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RewardItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Reward Item",
        "tags": [
          "rewards"
        ]
      }
    },
    "/api/v1/reward-items/{item_id}": {
      "delete": {
        "operationId": "delete_reward_item_api_v1_reward_items__item_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Reward Item",
        "tags": [
          "rewards"
        ]
      },
      "patch": {
        "operationId": "update_reward_item_api_v1_reward_items__item_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RewardItemUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RewardItemResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Reward Item",
        "tags": [
          "rewards"
        ]
      }
    },
    "/api/v1/reward-redemptions": {
      "get": {
        "operationId": "list_reward_redemptions_api_v1_reward_redemptions_get",
        "parameters": [
          {
            "in": "query",
            "name": "reader_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Reader Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RedemptionResponse"
                  },
                  "title": "Response List Reward Redemptions Api V1 Reward Redemptions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Reward Redemptions",
        "tags": [
          "rewards"
        ]
      },
      "post": {
        "operationId": "redeem_reward_api_v1_reward_redemptions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedemptionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RedemptionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Redeem Reward",
        "tags": [
          "rewards"
        ]
      }
    },
    "/api/v1/reward-redemptions/{redemption_id}": {
      "patch": {
        "operationId": "transition_redemption_api_v1_reward_redemptions__redemption_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "redemption_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Redemption Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedemptionAction"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RedemptionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Transition Redemption",
        "tags": [
          "rewards"
        ]
      }
    },
    "/api/v1/rewards/progress": {
      "get": {
        "operationId": "get_reward_progress_api_v1_rewards_progress_get",
        "parameters": [
          {
            "in": "query",
            "name": "reader_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Reader Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RewardProgressResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Reward Progress",
        "tags": [
          "rewards"
        ]
      }
    },
    "/api/v1/rewards/transactions": {
      "get": {
        "operationId": "list_reward_transactions_api_v1_rewards_transactions_get",
        "parameters": [
          {
            "in": "query",
            "name": "reader_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Reader Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RewardTransactionResponse"
                  },
                  "title": "Response List Reward Transactions Api V1 Rewards Transactions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Reward Transactions",
        "tags": [
          "rewards"
        ]
      }
    }
  }
}